@@ -1428,12 +1428,6 @@ public class OntGraphModelImpl extends ModelCom implements OntModel, OntEnhGraph
     }
 
     @Override
-    public OntGraphModelImpl add(Resource s, Property p, String lex, boolean wellFormed) {
-        super.add(s, p, lex, wellFormed);
-        return this;
-    }
-
-    @Override
     public OntGraphModelImpl add(Resource s, Property p, String lex, String lang) {
         super.add(s, p, lex, lang);
         return this;
@@ -1485,6 +1479,7 @@ public class OntGraphModelImpl extends ModelCom implements OntModel, OntEnhGraph
         return this;
     }
 
+    @SuppressWarnings("deprecation")
     @Override
     public OntGraphModelImpl write(Writer writer) {
         getBaseModel().write(writer);
